+++ tests/make_fixtures.py
@@ -236,7 +236,10 @@ build("cover.epub", {
 nav_doc = """<?xml version='1.0' encoding='utf-8'?>
 <html xmlns="http://www.w3.org/1999/xhtml" xmlns:epub="http://www.idpf.org/2007/ops"
       xml:lang="en">
-  <head><title>Contents</title></head>
+  <head><title>Contents</title>
+    <link rel="stylesheet" type="text/css" href="style.css"/>
+    <link rel="stylesheet" type="text/css" href="page_styles1.css"/>
+  </head>
   <body>
     <nav epub:type="toc"><ol>
       <li><a href="c0.xhtml">Cover</a></li>